pyproject.toml
[tool.poetry]
name = "web3-camera"
version = "0.1.0"
description = ""
authors = ["iain nash <me@iain.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
web3 = "6.11.0"
nft-storage = {git = "https://github.com/nftstorage/python-client.git"}
python-dotenv = "^1.0.0"
opencv-python = "^4.8.1.78"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
